Change the app's functioality and improve downloader
The application for now has been changed to only be a wrapper for
the youtube_dl python package. The downloader functionality now
automates the entire progress, like downloading, converting and
deleting old pre-converted files. Also does not include the video's
ID in the file name anymore. Prompts have been removed to allow
the automation, and command flags are no longer used. Furthermore,
enabled song requests to be stored in a file "songs.txt". Each line
counts as one request and can either be a YT search or a URI.